refactor :: [FRONTEND-71] api디렉토리 파일 구조 변경
