+++ src/api/user/signUp.ts
