+++ src/applications/utility/signUp/index.tsx
@@ -3,8 +3,9 @@ import Logo from '@/assets/windeath44.svg';
 import {useEffect, useState} from "react";
 import Button from "@/applications/components/button";
 import Inputs from "@/applications/components/inputs";
-import { useSignUp } from '@/api/user';
-import { useEmailValidation, useVerifyEmail } from '@/api/auth';
+import { useSignUp } from '@/api/user/signUp.ts';
+import { useEmailValidation} from '@/api/auth/emailValidationRequest.ts';
+import { useVerifyEmail } from '@/api/auth/verifyEmailCode.ts';
 type Props = {
     changeToLogIn: () => void;
 };
